+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.2.0-beta (In-Development)
+
+- Added update checker
+- Decreased font size to 10pt
+- Added version string parser support for rc (release candidates)
+- Fixed save/cancel button tooltips for preset items
+
 # 1.1.0-beta
 
 - General theme skeleton improved for all platforms